Move File upload to a separate controller
